added Microsoft.AspNetCore.Authentication.JwtBearer package for auth.

+++ src/entityFramework/entities/PaymentMethod.cs
@@ -8,7 +8,7 @@ public class PaymentMethod
     public required Guid UserId { get; set; }
     
     public required string Type { get; set; }
-    public required decimal CardNumber { get; set; }
+    public required double CardNumber { get; set; }
     public required string CardHolderName { get; set; }
     public required DateTime CardExpirationDate { get; set; }
     public required int CardCCV { get; set; }
